remove internal file
